git-svn-id: http://svn.vuze.com/public/plugins@28108 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ 3DView/com/aelitis/azureus/plugins/view3d/Plugin3D.java
@@ -180,10 +180,12 @@ public class Plugin3D implements Plugin {
 					break;
 				}
 				case UISWTViewEvent.TYPE_DESTROY:{
+					if ( _3D_view != null ){
 					
-					_3D_view.delete();
+						_3D_view.delete();
 					
-					_3D_view	= null;
+						_3D_view	= null;
+					}
 					
 					isCreated = false;
